Split command-bars.tsx into separate files for Assessments and Automated Checks
